server/requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
environs==14.3.0
fastapi==0.116.1
fastapi-mail==1.5.0
fastapi_cors==0.0.6
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
limits==5.5.0
MarkupSafe==3.0.2
marshmallow==4.0.0
openai==1.99.9
packaging==25.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
slowapi==0.1.9
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
